Move rtc implementation to core project and pass ILog as ctor parameter to it.
