+++ fellow/SRC/WinFellow.Core/VirtualHost/Core.cpp
@@ -3,7 +3,15 @@
 // For the time being this has to be a singleton
 Core _core = Core();
 
-Core::Core() : Registers(), RegisterUtility(Registers), Sound(nullptr), Uart(nullptr), Drivers(), Log(nullptr), Fileops(nullptr)
+Core::Core() : 
+  Registers(), 
+  RegisterUtility(Registers), 
+  Sound(nullptr), 
+  Uart(nullptr), 
+  RtcOkiMsm6242rs(nullptr), 
+  Drivers(), 
+  Log(nullptr), 
+  Fileops(nullptr)
 {
 }
