+++ fellow/SRC/WinFellow.Core/VirtualHost/Core.h
@@ -7,6 +7,7 @@
 #include "CustomChipset/Registers.h"
 #include "CustomChipset/RegisterUtility.h"
 #include "IO/Uart.h"
+#include "IO/RtcOkiMsm6242rs.h"
 
 class Core
 {
@@ -15,6 +16,7 @@ public:
   CustomChipset::RegisterUtility RegisterUtility;
   Sound* Sound;
   Uart* Uart;
+  RtcOkiMsm6242rs* RtcOkiMsm6242rs;
 
   Drivers Drivers;
   Service::ILog* Log;
