+++ fellow/SRC/WinFellow/C/rtc.cpp
@@ -19,18 +19,17 @@
 /* Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.          */
 /*=========================================================================*/
 
-#include "RtcOkiMsm6242rs.h"
+#include "IO/RtcOkiMsm6242rs.h"
 #include "FMEM.H"
 #include "FELLOW.H"
 #include "rtc.h"
 #include "VirtualHost/Core.h"
 
 bool rtc_enabled = false;
-RtcOkiMsm6242rs rtc;
 
 uint8_t rtcReadByte(uint32_t address)
 {
-  uint16_t result = rtc.read(address);
+  uint16_t result = _core.RtcOkiMsm6242rs->read(address);
   uint8_t byte_result = (uint8_t) ((address & 1) ? result : (result>>8));
 
 #ifdef RTC_LOG
@@ -42,7 +41,7 @@ uint8_t rtcReadByte(uint32_t address)
 
 uint16_t rtcReadWord(uint32_t address)
 {
-  uint16_t result = rtc.read(address);
+  uint16_t result = _core.RtcOkiMsm6242rs->read(address);
 
 #ifdef RTC_LOG
   _core.Log->AddLog("RTC Word Read: %.8X, returned %.4X\n", address, result);
@@ -53,8 +52,8 @@ uint16_t rtcReadWord(uint32_t address)
 
 uint32_t rtcReadLong(uint32_t address)
 {
-  uint32_t w1 = (uint32_t) rtc.read(address);
-  uint32_t w2 = (uint32_t) rtc.read(address+2);
+  uint32_t w1 = (uint32_t)_core.RtcOkiMsm6242rs->read(address);
+  uint32_t w2 = (uint32_t)_core.RtcOkiMsm6242rs->read(address+2);
   uint32_t result = (w1 << 16) | w2;
 
 #ifdef RTC_LOG
@@ -66,7 +65,7 @@ uint32_t rtcReadLong(uint32_t address)
 
 void rtcWriteByte(uint8_t data, uint32_t address)
 {
-  rtc.write(data, address);
+  _core.RtcOkiMsm6242rs->write(data, address);
 
 #ifdef RTC_LOG
   _core.Log->AddLog("RTC Byte Write: %.8X %.2X\n", address, data);
@@ -75,7 +74,7 @@ void rtcWriteByte(uint8_t data, uint32_t address)
 
 void rtcWriteWord(uint16_t data, uint32_t address)
 {
-  rtc.write(data, address);
+  _core.RtcOkiMsm6242rs->write(data, address);
 
 #ifdef RTC_LOG
   _core.Log->AddLog("RTC Word Write: %.8X %.4X\n", address, data);
@@ -84,8 +83,8 @@ void rtcWriteWord(uint16_t data, uint32_t address)
 
 void rtcWriteLong(uint32_t data, uint32_t address)
 {
-  rtc.write(data, address);
-  rtc.write(data, address + 2);
+  _core.RtcOkiMsm6242rs->write(data, address);
+  _core.RtcOkiMsm6242rs->write(data, address + 2);
 
 #ifdef RTC_LOG
   _core.Log->AddLog("RTC Long Write: %.8X %.8X\n", address, data);
